Water 1st Iteration
It's not perfect, but now there's a seamless basic water material
material.

Note: You'll notice that the "pool" the water is kept in does not change
colour when changing world state. That is because that is a brush and
not a static mesh actor, therefore the current method does not cover it.
But since our actual levels probably won't be done with brushes, it's
not really an issue.