Some minor changes
Had to remove all the grass was causing major performance issues even on
my GTX1060. Especially with Light building.

Adjusted Foliage material to two sided as some grass and bushes
disapeared based on certain angles.

Added platforms in the beggining area to jump on.

Will need to use Foliage brushes sparsely. Be conservative with where we
plonk it all.